rename ProfileSummary and removed the extra div.
