+++ src/client/components/ProfileSummary/ProfileSummary.css
@@ -1,0 +1,16 @@
+.profile-Image {
+  height: 35px;
+  width: 35px;
+  border-radius: 50%;
+  padding-right: 10px;
+}
+.profile-container {
+  border-bottom: solid 2px #7cfc00;
+  width: 20%;
+  display: flex;
+  padding-bottom: 10px;
+}
+.profile-Name {
+  font-family: "Open Sans", sans-serif;
+  padding-top: 7%;
+}
