--- src/client/components/ProfileSummery/ProfileSummery.css
@@ -1,17 +1,0 @@
-.profileImage {
-    height: 35px;
-    width: 35px;
-    border-radius: 50%;
-    padding-right:10px;
-  }
-  .profile-container {
-    border-bottom: solid 2px #7cfc00;
-    width: 20%;
-    display: flex;
-    padding-bottom: 10px;
-  }
-  .profileName{
-      font-family: "Open Sans", sans-serif;
-      padding-top: 7%;
-  }
-
