+++ src/client/stories/index.js
@@ -2,4 +2,4 @@ import React from "react";
 import { storiesOf } from "@storybook/react";
 import { Button } from "@storybook/react/demo";
 import { SidebarMenu } from "../components/SidebarMenu/SidebarMenu.story";
-import { ProfileSummery } from "../components/ProfileSummery/ProfileSummery.story";
+import {ProfileSummary} from "../components/ProfileSummary/ProfileSummary.story"
